fix code review comments
fix code review comments
