+++ projects/core/src/auth/user-auth/facade/auth.service.ts
@@ -25,8 +25,8 @@ import { AuthStorageService } from '../services/auth-storage.service';
 import { OAuthLibWrapperService } from '../services/oauth-lib-wrapper.service';
 import { AuthActions } from '../store/actions/index';
 import { UserIdService } from './user-id.service';
-import { LoginModeIndicatorService } from './loginModeIndicatorService';
-import { FeatureToggles } from '@spartacus/core';
+import { FeatureToggles, WindowRef } from '@spartacus/core';
+import { Location } from '@angular/common';
 
 /**
  * Auth service for normal user authentication.
@@ -49,7 +49,9 @@ export class AuthService {
    */
   logoutInProgress$: Observable<boolean> = new BehaviorSubject<boolean>(false);
 
-  private loginModeIndicatorService = inject(LoginModeIndicatorService);
+  protected location = inject(Location);
+  protected winRef = inject(WindowRef);
+
   protected csrfToken$ = this.crossSiteRequestForgeryService
     .getCsrfToken()
     .pipe(shareReplay({ bufferSize: 1, refCount: true }));
@@ -229,14 +231,46 @@ export class AuthService {
     (this.logoutInProgress$ as BehaviorSubject<boolean>).next(progress);
   }
 
+  /**
+   * Indicates whether the ASM module is enabled.
+   */
+  isAsmEnabled(): boolean {
+    if (this.isLaunched() && !this.isUsedBefore() && this.winRef.localStorage) {
+      this.winRef.localStorage.setItem('asm_enabled', 'true');
+    }
+    return this.isLaunched() || this.isUsedBefore() || this.isEmulateInURL();
+  }
+
+  /**
+   * Indicates whether ASM is launched through the URL,
+   * using the asm flag in the URL.
+   */
+  protected isLaunched(): boolean {
+    const params = this.location.path().split('?')[1];
+    return !!params && params.split('&').includes('asm=true');
+  }
+
+  /**
+   * check whether try to emulate customer from deeplink
+   * */
+  protected isEmulateInURL(): boolean {
+    return this.location.path().indexOf('assisted-service/emulate?') > 0;
+  }
+
+  /**
+   * Evaluates local storage where we persist the usage of ASM.
+   */
+  protected isUsedBefore(): boolean {
+    if (this.winRef.localStorage) {
+      return this.winRef.localStorage.getItem('asm_enabled') === 'true';
+    } else {
+      return false;
+    }
+  }
+
   public refreshAuthConfig() {
-    if (
-      this.loginModeIndicatorService.isEnabled() &&
-      this.authorizationCodeFlowByDefault
-    ) {
-      this.oAuthLibWrapperService.changeAsmAuthConfigClientId(
-        'mobile_android_asm'
-      );
+    if (this.isAsmEnabled() && this.authorizationCodeFlowByDefault) {
+      this.oAuthLibWrapperService.changeAuthConfigClientId('asm_client');
     } else {
       this.oAuthLibWrapperService.refreshAuthConfig();
     }
