+++ projects/core/src/auth/user-auth/services/oauth-lib-wrapper.service.ts
@@ -196,7 +196,7 @@ export class OAuthLibWrapperService {
     this.initialize();
   }
 
-  public changeAsmAuthConfigClientId(clientId: string) {
+  public changeAuthConfigClientId(clientId: string) {
     this.changeClientWhenInitialize(clientId);
   }
 }
